+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 Changelog
 =========================
 
+### v2.6.1 Release
+#### New Features
+- New (Spot Sub-Account)
+    - `/account/sub-account/main/v1/sub-to-main` Sub-Account to Main-Account, For Main Account (SIGNED)
+    - `/account/sub-account/sub/v1/sub-to-main` Sub-Account to Main-Account, For Sub Account (SIGNED)
+    - `/account/sub-account/main/v1/main-to-sub` Main-Account to Sub-Account, For Main Account (SIGNED)
+    - `/account/sub-account/main/v1/sub-to-sub` Sub-Account to Sub-Account, For Main Account (SIGNED)
+    - `/account/sub-account/main/v1/transfer-list` Get Sub-Account Transfer History, For Main Account (KEYED)
+    - `/account/sub-account/v1/transfer-history` Get Account Transfer History, For Main and Sub Account (KEYED)
+    - `/account/sub-account/main/v1/wallet` Get Sub-Account Spot Wallet Balance, For Main Account (KEYED)
+    - `/account/sub-account/main/v1/subaccount-list` Get Sub-Account List, For Main Account (KEYED)
+- New (Futures Sub-Account)
+    - `/account/contract/sub-account/main/v1/sub-to-main` Sub-Account to Main-Account, For Main Account (SIGNED)
+    - `/account/contract/sub-account/sub/v1/sub-to-main` Sub-Account to Main-Account, For Sub Account (SIGNED)
+    - `/account/contract/sub-account/main/v1/main-to-sub` Main-Account to Sub-Account, For Main Account (SIGNED)
+    - `/account/contract/sub-account/main/v1/transfer-list` Get Sub-Account Transfer History, For Main Account (KEYED)
+    - `/account/contract/sub-account/v1/transfer-history` Get Account Transfer History, For Main and Sub Account (KEYED)
+    - `/account/contract/sub-account/main/v1/wallet` Get Sub-Account Futures Wallet Balance, For Main Account (KEYED)
+
+
+---
+
 ### v2.6.0 Release
 #### New Features
 - New